Array destructuring + move map style
