+++ packages/Auxilium-App/src/screens/Main/Home/Overview/components/NearbyMap.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, Dimensions } from 'react-native';
 import { MapView, Region } from 'expo';
-import mapStyle from '../mapStyle';
+import mapStyle from './mapStyle';
 import MapMarker from './MapMarker';
 
 const { width, height } = Dimensions.get('window');
@@ -15,17 +15,14 @@ interface NearbyMapProps {
 
 const renderMarkers = (emergencies: Emergency[]) => {
 	if (!emergencies) return null;
-	return emergencies.map((emergency: Emergency, index: number) => (
-		<MapView.Marker
-			key={index}
-			coordinate={{
-				longitude: emergency.location.coordinates[0],
-				latitude: emergency.location.coordinates[1]
-			}}
-		>
-			<MapMarker size={16} />
-		</MapView.Marker>
-	));
+	return emergencies.map((emergency: Emergency, index: number) => {
+		const [longitude, latitude] = emergency.location.coordinates;
+		return (
+			<MapView.Marker key={index} coordinate={{ longitude, latitude }}>
+				<MapMarker size={16} />
+			</MapView.Marker>
+		);
+	});
 };
 
 const NearbyMap = ({
