+++ packages/Auxilium-App/src/screens/Main/Home/Overview/components/mapStyle.ts
